AG-16239 - Add reducer + processor incremental support.
