+++ packages/ag-charts-community/src/chart/data/data-model/reducers/bandedReducer.ts
@@ -1,4 +1,5 @@
 import type { BandedDomainConfig } from '../../dataDomain';
+import type { ReducerOutputPropertyDefinition } from '../../dataModelTypes';
 import { type BandLike, BandedStructure } from '../utils/bandedStructure';
 
 export interface ReducerBand extends BandLike {
@@ -13,14 +14,23 @@ export interface BandedReducerStats extends Record<string, number> {
     cacheHits: number;
 }
 
+export interface ReducerContext {
+    rawData: unknown[];
+    keyColumns: unknown[][];
+    keysParam: unknown[];
+}
+
 /**
  * Band-based structure for reducer aggregations.
  * Each band maintains a cached result for efficient incremental updates.
  *
  * This class extends BandedStructure with reducer-specific functionality:
  * - Caching aggregation results per band
+ * - Evaluating reducers across bands with overlap support
+ * - Combining band results into final aggregated values
  * - Tracking scan ratios for performance metrics
- * - Providing statistics about cache hits and dirty bands
+ *
+ * Symmetrical to BandedDomain which handles domain aggregation.
  */
 export class BandedReducer extends BandedStructure<ReducerBand> {
     private lastDirtyBandCount: number = 0;
@@ -52,10 +62,85 @@ export class BandedReducer extends BandedStructure<ReducerBand> {
     }
 
     /**
-     * Gets the array of bands for direct access by reducers.
+     * Gets the array of bands for direct access.
+     * @deprecated Use evaluateFromData() and getResult() instead for symmetry with BandedDomain.
      */
     getBands(): ReducerBand[] {
         return this.bands;
+    }
+
+    /**
+     * Evaluates a reducer across all bands, reusing cached results for clean bands.
+     * Symmetrical to BandedDomain.extendBandsFromData().
+     *
+     * @param def Reducer definition with reducer function, initial value, and overlap settings
+     * @param context Reducer context containing raw data and key columns
+     * @param reuseCleanBands Whether to reuse cached results for clean bands (default: false)
+     */
+    evaluateFromData(
+        def: ReducerOutputPropertyDefinition,
+        context: ReducerContext,
+        reuseCleanBands: boolean = false
+    ): void {
+        const reducerFn = def.reducer();
+
+        for (const band of this.bands) {
+            if (reuseCleanBands && !band.isDirty) {
+                continue; // Keep cached result
+            }
+
+            // Handle overlap for interval reducers (e.g., min/max interval calculations)
+            const startIndex =
+                def.needsOverlap && band.startIndex > 0 ? Math.max(0, band.startIndex - 1) : band.startIndex;
+
+            const result = this.evaluateRange(def, reducerFn, context, startIndex, band.endIndex);
+            band.cachedResult = result;
+            band.isDirty = false;
+        }
+    }
+
+    /**
+     * Combines all band results to get the final aggregated value.
+     * Symmetrical to BandedDomain.getDomain().
+     *
+     * @param def Reducer definition with combineResults function
+     * @returns Combined result from all bands
+     */
+    getResult(def: ReducerOutputPropertyDefinition): unknown {
+        const bandResults = this.bands.map((band) => band.cachedResult);
+        return def.combineResults!(bandResults as any[]);
+    }
+
+    /**
+     * Evaluates a reducer over a specific range of data indices.
+     * Symmetrical to BandedDomain's band scanning loop in extendBandsFromData().
+     *
+     * @param def Reducer definition with initial value
+     * @param reducer Reducer function to apply
+     * @param context Reducer context with data and keys
+     * @param startIndex Starting index (inclusive)
+     * @param endIndex Ending index (exclusive)
+     * @returns Accumulated reducer result for the range
+     */
+    private evaluateRange(
+        def: ReducerOutputPropertyDefinition,
+        reducer: ReturnType<ReducerOutputPropertyDefinition['reducer']>,
+        context: ReducerContext,
+        startIndex: number,
+        endIndex: number
+    ): unknown {
+        let accValue: any = def.initialValue;
+        const { keyColumns, keysParam, rawData } = context;
+        const clampedEnd = Math.min(endIndex, rawData.length);
+
+        for (let datumIndex = startIndex; datumIndex < clampedEnd; datumIndex += 1) {
+            for (let keyIdx = 0; keyIdx < keysParam.length; keyIdx++) {
+                keysParam[keyIdx] = keyColumns[keyIdx]?.[datumIndex];
+            }
+            accValue = reducer(accValue, keysParam);
+        }
+
+        return accValue;
     }
 
     /**
